Tests/LibWeb/Text/expected/css/CSSStyleDeclaration-modify-computed.txt

Calling setProperty() on a computed CSSStyleDeclaration throws error of type: NoModificationAllowedError
Calling removeProperty() on a computed CSSStyleDeclaration throws error of type: NoModificationAllowedError
